Make bind-fields return a form, not a function.
Previously, `bind-fields` would transform `form` as a setup step
(i.e. it would transform it and then return a function that would give
the 'pre-transformed' `form`).

This had the unfortunate consequence that forms would not be properly
updated when the `doc` atom changed.

This commit fixes this issue by making `bind-fields` return the form
itself. This way, reagent will make sure that the forms are updated
properly.